backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Polkadot
substrate-interface

# Cache & Queue
redis==5.0.1
celery==5.3.4

# Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSocket
websockets==12.0

# HTTP Client
httpx==0.25.2

# QR Code
qrcode==7.4.2
pillow==10.1.0

# Utilities
python-dotenv==1.0.0

